refactor: move the mobile bank transaction views into the domain
BankTransactionsMobileList (12 private parts nested, plus purpose.ts) and
BankTransactionsMobileCategorySelection.

The category selection stays a sibling rather than nesting inside the mobile
list, even though BankTransactionsMobileListBusinessForm is its only caller: it
is five files with its own item component, stylesheet pair, and utils, and the
SKILL's test is whether another component could render it, not how many do. A
category-selection drawer is not intrinsically list-private.

BankTransactionsMobileList/constants.ts -> personalCategories.ts. It holds only
PersonalStableName and LegacyPersonalCategories, and a bare constants.ts inside a
component directory invites drift toward the domain root's.

Note the one non-mechanical fix in this commit: running eslint --fix while the
renamed import was momentarily unresolved made no-unnecessary-type-assertion
delete a load-bearing `as LegacyPersonalCategories` cast, since the module read
as any. Restored.
