+++ src/components/BankTransactions/BankTransactions.tsx
@@ -30,10 +30,10 @@ import { BankTransactionsStringOverridesProvider } from '@contexts/BankTransacti
 import { CategorizationRulesContext, CategorizationRulesProvider } from '@contexts/CategorizationRulesContext/CategorizationRulesContext'
 import { InAppLinkProvider, type LinkingMetadata } from '@contexts/InAppLinkContext'
 import { Container } from '@blocks/Layout/Container/Container'
-import { BankTransactionsMobileList } from '@components/BankTransactionsMobileList/BankTransactionsMobileList'
 import { ErrorBoundary } from '@components/utility/ErrorBoundary/ErrorBoundary'
 import { BankTransactionsHeader } from '@features/bankTransactions/BankTransactionsHeader/BankTransactionsHeader'
 import { BankTransactionsList } from '@features/bankTransactions/BankTransactionsList/BankTransactionsList'
+import { BankTransactionsMobileList } from '@features/bankTransactions/BankTransactionsMobileList/BankTransactionsMobileList'
 import { BankTransactionsTable } from '@features/bankTransactions/BankTransactionsTable/BankTransactionsTable'
 import { BankTransactionsTableContent, type MobileComponentType } from '@features/bankTransactions/constants'
 import { type BankTransactionsStringOverrides } from '@features/bankTransactions/types'
