+++ src/components/features/bankTransactions/BankTransactionsMobileCategorySelection/bankTransactionsMobileCategorySelection.scss
