+++ src/components/features/bankTransactions/BankTransactionsMobileCategorySelection/bankTransactionsMobileCategorySelectionItem.scss
