+++ src/components/features/bankTransactions/BankTransactionsMobileList/BankTransactionsMobileListItemFooter.tsx
