+++ src/components/features/bankTransactions/BankTransactionsMobileList/bankTransactionsMobileList.scss
