+++ src/components/features/bankTransactions/BankTransactionsMobileList/bankTransactionsMobileListItem.scss
