+++ src/components/features/bankTransactions/BankTransactionsMobileList/bankTransactionsMobileListSplitForm.scss
