+++ src/components/features/bankTransactions/BankTransactionsMobileList/personalCategories.ts
