+++ src/components/features/bankTransactions/BankTransactionsMobileList/purpose.ts
